Initial Commit. All seems to be working...?
EditPage.jsx and GoogleActions.js need to be cleaned up and shortened...

+++ client/homebrew/pages/userPage/brewItem/brewItem.jsx
@@ -6,6 +6,8 @@ const cx    = require('classnames');
 const moment = require('moment');
 const request = require('superagent');
 
+const googleDriveIcon = require('../../../googleDrive.png');
+
 const BrewItem = createClass({
 	getDefaultProps : function() {
 		return {
@@ -27,11 +29,19 @@ const BrewItem = createClass({
 			if(!confirm('Are you REALLY sure? You will lose editor access to this document.')) return;
 		}
 
-		request.delete(`/api/${this.props.brew.editId}`)
-			.send()
-			.end(function(err, res){
-				location.reload();
-			});
+		if(this.props.brew.googleId) {
+			request.get(`/api/removeGoogle/${this.props.brew.googleId}${this.props.brew.editId}`)
+				.send()
+				.end(function(err, res){
+					location.reload();
+				});
+		} else {
+			request.delete(`/api/${this.props.brew.editId}`)
+				.send()
+				.end(function(err, res){
+					location.reload();
+				});
+		}
 	},
 
 	renderDeleteBrewLink : function(){
@@ -41,12 +51,39 @@ const BrewItem = createClass({
 			<i className='fa fa-trash' />
 		</a>;
 	},
+
 	renderEditLink : function(){
 		if(!this.props.brew.editId) return;
 
-		return <a href={`/edit/${this.props.brew.editId}`} target='_blank' rel='noopener noreferrer'>
+		let editLink = this.props.brew.editId;
+		if(this.props.brew.googleId) {
+			editLink = this.props.brew.googleId + editLink;
+		}
+
+		return <a href={`/edit/${editLink}`} target='_blank' rel='noopener noreferrer'>
 			<i className='fa fa-pencil' />
 		</a>;
+	},
+
+	renderShareLink : function(){
+		if(!this.props.brew.shareId) return;
+
+		let shareLink = this.props.brew.shareId;
+		if(this.props.brew.googleId) {
+			shareLink = this.props.brew.googleId + shareLink;
+		}
+
+		return <a href={`/share/${shareLink}`} target='_blank' rel='noopener noreferrer'>
+			<i className='fa fa-share-alt' />
+		</a>;
+	},
+
+	renderGoogleDriveIcon : function(){
+		if(!this.props.brew.gDrive) return;
+
+		return <span>
+			<img className='googleDriveIcon' src={googleDriveIcon} alt='googleDriveIcon' />
+		</span>;
 	},
 
 	render : function(){
@@ -66,12 +103,11 @@ const BrewItem = createClass({
 				<span>
 					<i className='fa fa-refresh' /> {moment(brew.updatedAt).fromNow()}
 				</span>
+				{this.renderGoogleDriveIcon()}
 			</div>
 
 			<div className='links'>
-				<a href={`/share/${brew.shareId}`} target='_blank' rel='noopener noreferrer'>
-					<i className='fa fa-share-alt' />
-				</a>
+				{this.renderShareLink()}
 				{this.renderEditLink()}
 				{this.renderDeleteBrewLink()}
 			</div>
